Move token config to its own file
